requirements.txt
# Core dependencies - PINNED VERSIONS for stability
requests==2.31.0
pdfplumber==0.10.3
python-dotenv==1.0.0
watchdog==3.0.0
lxml==4.9.0
alegra-python==0.1.3
pytesseract==0.3.10
Pillow==10.0.0
opencv-python==4.8.0

# Configuration and validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Async processing
aiohttp==3.9.1
asyncio==3.4.3

# Caching and queues
redis==5.0.1
celery==5.3.4